+++ src/zope/container/ordered.py
@@ -160,6 +160,11 @@ class OrderedContainer(Persistent, Contained):
 
     has_key = __contains__
 
+    def _setitemf(self, key, value):
+        if key not in self._data:
+            self._order.append(key)
+        self._data[key] = value
+
     def __setitem__(self, key, object):
         """ See `IOrderedContainer`.
 
@@ -182,26 +187,11 @@ class OrderedContainer(Persistent, Contained):
         >>> oc._order
         ['foo', 'baz']
         """
-
-        existed = key in self._data
-
-        # We have to first update the order, so that the item is available,
-        # otherwise most API functions will lie about their available values
-        # when an event subscriber tries to do something with the container.
-        if not existed:
-            # XXX: This is not using the decoded key that setitem will use.
-            # In the case of a bytes key. This is bad on Python 3. But also not tested.
-            # See https://github.com/zopefoundation/zope.container/issues/17
-            self._order.append(key)
-
-        # This function creates a lot of events that other code listens to.
-        try:
-            setitem(self, self._data.__setitem__, key, object)
-        except Exception:
-            if not existed:
-                self._order.remove(key)
-            raise
-
+        # This function creates a lot of events that other code
+        # listens to. None of them are fired (notified) though, before
+        # _setitemf is called. At that point we need to be sure that
+        # we have the key in _order too.
+        setitem(self, self._setitemf, key, object)
         return key
 
     def __delitem__(self, key):